chore: Add example projects
