+++ examples/async-views/README.md
@@ -1,0 +1,10 @@
+# Async views
+
+Example code for setting up correlation IDs for async views.
+
+To demo, run
+
+- `python manage.py runserver 8080`, or
+- `uvicorn src.asgi:application`
+
+and go to http://127.0.0.1:8080.
