+++ examples/async-views/src/urls.py
@@ -1,0 +1,5 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [path('', views.index_view, name='index')]
