+++ examples/celery/src/urls.py
@@ -1,0 +1,1 @@
+urlpatterns = []
