+++ examples/ignore-url/src/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('ignored', views.ignored_view, name='ignored'),
+]
