+++ examples/sync-views/README.md
@@ -1,0 +1,10 @@
+# Sync views
+
+Example code for setting up correlation IDs for sync views.
+
+To demo, run
+
+- `python manage.py runserver 8080`, or
+- `wsgi src.wsgi`
+
+and go to http://127.0.0.1:8080
