+++ examples/sync-views/src/services.py
@@ -1,0 +1,7 @@
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+def some_sync_function() -> None:
+    logger.info('Doing some work')
